refactor: CheckEventTarget.js에서 vendingMachineManageMenu 관련내용 파일 분리
- CheckVendingMachineManageMenu.js로 파일 분리
